refactor: General 이동 전략과 Guard 이동 전략 추상화
