+++ src/main/java/domain/place/moveStrategy/GuardMoveStrategy.java
@@ -1,27 +1,4 @@
 package domain.place.moveStrategy;
 
-import domain.board.BoardView;
-import domain.position.Position;
-import java.util.List;
-
-public class GuardMoveStrategy implements MoveStrategy {
-
-    private static final List<Direction> ORTHOGONAL_DIRECTIONS = List.of(
-            Direction.DOWN, Direction.LEFT, Direction.RIGHT, Direction.TOP
-    );
-
-    @Override
-    public boolean canMove(BoardView board, Position from, Position to) {
-        return canReachAdjacentPosition(from, to);
-    }
-
-    private boolean canReachAdjacentPosition(Position from, Position to) {
-        int currentRow = from.getRow();
-        int currentColumn = from.getColumn();
-
-        return ORTHOGONAL_DIRECTIONS.stream()
-                .filter(d -> Position.isNotOutOfBounds(currentRow + d.getRow(), currentColumn + d.getColumn()))
-                .map(from::move)
-                .anyMatch(to::equals);
-    }
+public class GuardMoveStrategy extends OrthogonalOneStepMoveStrategy {
 }
