+++ src/main/java/domain/place/moveStrategy/OrthogonalOneStepMoveStrategy.java
@@ -1,0 +1,27 @@
+package domain.place.moveStrategy;
+
+import domain.board.BoardView;
+import domain.position.Position;
+import java.util.List;
+
+public abstract class OrthogonalOneStepMoveStrategy implements MoveStrategy {
+
+    private static final List<Direction> ORTHOGONAL_DIRECTIONS = List.of(
+            Direction.DOWN, Direction.LEFT, Direction.RIGHT, Direction.TOP
+    );
+
+    @Override
+    public boolean canMove(BoardView board, Position from, Position to) {
+        return canReachAdjacentPosition(from, to);
+    }
+
+    protected boolean canReachAdjacentPosition(Position from, Position to) {
+        int currentRow = from.getRow();
+        int currentColumn = from.getColumn();
+
+        return ORTHOGONAL_DIRECTIONS.stream()
+                .filter(d -> Position.isNotOutOfBounds(currentRow + d.getRow(), currentColumn + d.getColumn()))
+                .map(from::move)
+                .anyMatch(to::equals);
+    }
+}
